Incorporared netwrok structs for realtTimeData

+++ models/network/weatherData/realTimeData.go
@@ -14,7 +14,7 @@ type RealTimeData struct {
 	Current struct {
 		LastUpdatedEpoch int64   `json:"last_updated_epoch"`
 		LastUpdated      string  `json:"last_updated"`
-		TempC            int     `json:"temp_c"`
+		TempC            float64 `json:"temp_c"`
 		TempF            float64 `json:"temp_f"`
 		IsDay            int     `json:"is_day"`
 		Condition        struct {
@@ -26,17 +26,17 @@ type RealTimeData struct {
 		WindKph    float64 `json:"wind_kph"`
 		WindDegree int     `json:"wind_degree"`
 		WindDir    string  `json:"wind_dir"`
-		PressureMb int     `json:"pressure_mb"`
+		PressureMb float64 `json:"pressure_mb"`
 		PressureIn float64 `json:"pressure_in"`
-		PrecipMm   int     `json:"precip_mm"`
+		PrecipMm   float64 `json:"precip_mm"`
 		PrecipIn   float64 `json:"precip_in"`
 		Humidity   int     `json:"humidity"`
 		Cloud      int     `json:"cloud"`
 		FeelslikeC float64 `json:"feelslike_c"`
 		FeelslikeF float64 `json:"feelslike_f"`
-		VisKm      int     `json:"vis_km"`
-		VisMiles   int     `json:"vis_miles"`
-		UV         int     `json:"uv"`
+		VisKm      float64 `json:"vis_km"`
+		VisMiles   float64 `json:"vis_miles"`
+		UV         float64 `json:"uv"`
 		GustMph    float64 `json:"gust_mph"`
 		GustKph    float64 `json:"gust_kph"`
 	} `json:"current"`
